+++ src/currency/index.ts
@@ -1,3 +1,5 @@
 export * from './Currency';
 export * from './iso4217';
 export * from './precision';
+export * from './registry';
+