+++ src/currency/precision.ts
@@ -1,4 +1,4 @@
-import { Currency } from './Currency';
+import { Currency } from "./Currency";
 
 export function getMinorUnitExponent(currency: Currency): bigint {
   return 10n ** BigInt(currency.decimals);