+++ src/errors/RoundingRequiredError.ts
@@ -1,7 +1,7 @@
-import { MonetraError } from './BaseError';
+import { MonetraError } from "./BaseError";
 
 export class RoundingRequiredError extends MonetraError {
   constructor() {
-    super('Rounding is required for this operation but was not provided.');
+    super("Rounding is required for this operation but was not provided.");
   }
 }