+++ src/money/index.ts
@@ -1,2 +1,4 @@
-export * from './Money';
-export * from './guards';
+export * from "./Money";
+export * from "./guards";
+export * from "./Converter";
+export * from "./MoneyBag";